creo componente de panalla de formulario y componente para pasos a seguir
