+++ src/components/Timelines/StepsDashboard.tsx
@@ -1,0 +1,26 @@
+import { Button, Timeline } from 'flowbite-react';
+import { HiArrowNarrowRight } from 'react-icons/hi';
+
+const StepsDashboard = () => {
+  return (
+    <Timeline>
+      <Timeline.Item>
+        <Timeline.Point />
+        <Timeline.Content>
+          <Timeline.Time>February 2022</Timeline.Time>
+          <Timeline.Title>Application UI code in Tailwind CSS</Timeline.Title>
+          <Timeline.Body>
+            Get access to over 20+ pages including a dashboard layout, charts,
+            kanban board, calendar, and pre-order E-commerce & Marketing pages.
+          </Timeline.Body>
+          <Button color="primary">
+            Learn More
+            <HiArrowNarrowRight className="ml-2 h-3 w-3" />
+          </Button>
+        </Timeline.Content>
+      </Timeline.Item>
+    </Timeline>
+  );
+};
+
+export { StepsDashboard };
